✨ feat: 랜덤 숫자 생성 후 중복체크

+++ src/main/kotlin/Main.kt
@@ -1,11 +1,26 @@
 package org.example
 
 import inline.input
+import kotlin.random.Random
 
 fun main() {
     println("< 게임을 시작합니다 >")
 
-    val num = input("숫자를 입력하세요").toInt()
+
+    val num = input("숫자를 입력하세요")
+
+
+//    val answerNum = "110"
+
+
+
+
+
+
+
+
+
+
 
 
 }
